refactor(ffi-bridge): split error.rs into error/ directory module
Pure refactor: 997-line error.rs becomes a 4-file directory module per
issue #36. Public surface (crate::error::FfiUnlockError /
FfiVaultError) is unchanged; both enums plus their From impls move to
sibling submodules with byte-identical Display strings.

- error/mod.rs (72)        — module docs + re-exports
- error/unlock.rs (316)    — FfiUnlockError + From<core::UnlockError>
- error/vault.rs (524)     — FfiVaultError + From<core::VaultError>
- error/conversions.rs (122) — From<FfiUnlockError> for FfiVaultError +
                                byte-identical-mirror tripwire test

Test counts unchanged: 599 passed + 9 ignored.
